pyproject.toml
[project]
name = "team-finder"
version = "0.1.0"
description = "API for a team finding application"
requires-python = ">=3.13"
dependencies = [
    "aiobcrypt>=3.2.0.post1",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.12",
    "logfire[fastapi,sqlalchemy]>=3.14.0",
    "orjson>=3.10.16",
    "pydantic>=2.11.3",
    "pyjwt[crypto]>=2.10.1",
    "sqlmodel>=0.0.24",
]

[dependency-groups]
dev = [
    "ruff>=0.11.6",
]


[tool.logfire]
ignore_no_config = true


[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
select = [
    "F", "E4", "E7", "E9",  # default (Pyflakes + pycodestyle subset)
    "S",                    # flake8-bandit (security)
    "B",                    # flake8-bugbear (common bug causes)
    "A",                    # flake8-builtins (builtin name shadowing)
    "C4",                   # flake8-comprehensions
    "ANN",                  # flake8-annotations
    "ARG",                  # flake8-unused-arguments
    "ASYNC",                # flake8-async
    "FBT",                  # flake8-boolean-trap (disallow boolean positional arguments)
    "RET",                  # flake8-return
    "SIM",                  # flake8-simplify
    "PTH",                  # flake8-use-pathlib
    "I",                    # isort (imports)
    "N",                    # pep8-naming
    "UP",                   # pyupgrade (disallow deprecated language features)
    "FURB",                 # refurb (use new language features)
    "RUF",                  # ruff-specific
    "PT",                   # pytest
    "FAST",                 # FastAPI
]
fixable = ["ALL"]

ignore = [
    "ANN002",  # allow missing annotation for *args
    "ANN003",  # allow missing annotation for **kwargs
    "RUF001",  # do not warn about ambiguous characters in strings
    "RUF002",  # do not warn about ambiguous characters in docstrings
]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.ruff.format]
line-ending = "lf"
skip-magic-trailing-comma = true
